chore: Remove Discord
The idea here is to move to two separate repositories,
one for Discord and one for IRC. This allows us to expand
support to Matrix or other protocols in the future more
easily.

There were several drawbacks from treating Discord and IRC
the same, such as the case where only half of the client
is suffering connection issues.